chore(suite): move graph worker to support folder
